requirements.txt
BeautifulSoup4
EbookLib
dulwich>=0.23.1
dublib[telebot]==0.24.3
pillow